# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flareo v1 API",
    "version": "0.2.0",
    "description": "Public API for the Flareo container supply chain platform. Covers image verification against Sigstore, module catalog reads, health checks, and identity lookups. The CLI consumes the same endpoints described here.\n\nAuthentication is optional for read endpoints and `/verify`, and either a NextAuth session cookie or an `Authorization: Bearer fla_<token>` header is accepted where required.\n\nRate limits: 60/hour/IP for anonymous verify calls, 600/hour/user for authenticated, 300/hour/key on catalog reads. Exceeded limits return 429 with `Retry-After` set."
  },
  "servers": [
    { "url": "https://flareo.dev", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "paths": {
...
    "/api/v1/modules/{slug}": {
      "get": {
        "summary": "Get module by slug",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Module detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModuleDetail" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/whoami": {
      "get": {
        "summary": "Get authenticated user",
        "description": "Accepts either a NextAuth session cookie or an `Authorization: Bearer fla_<token>` header.",
        "responses": {
          "200": { "description": "Authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WhoamiResult" } } } },
          "401": { "description": "No valid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "fla_<token>"
      }
    },
    "schemas": {
...
      "ModuleSummary": {
        "type": "object",
        "properties": {
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "version": { "type": "string" },
          "author": { "type": "string" },
          "description": { "type": "string" },
          "category": { "type": "string" },
          "status": { "type": "string", "enum": ["verified", "pending", "failing"] },
          "slsa": { "type": "string" },
          "trust": { "type": "integer" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "cves": { "$ref": "#/components/schemas/CveCounts" },
          "deploys": { "type": "integer" },
          "digest": { "type": "string" },
          "previewable": { "type": "boolean" },
          "lastRebuiltAt": { "type": ["string", "null"], "format": "date-time" }
        }
      },
      "ModuleDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/ModuleSummary" },
          {
            "type": "object",
            "properties": {
              "trustBreakdown": {
                "type": "object",
                "properties": {
                  "vulns": { "type": "integer" },
                  "slsa": { "type": "integer" },
                  "signature": { "type": "integer" },
                  "sbom": { "type": "integer" }
                }
              },
              "size": { "type": "string" },
              "imageRef": { "type": ["string", "null"] },
              "upstreamRef": { "type": ["string", "null"] },
              "upstreamDigest": { "type": ["string", "null"] },
              "sbomUrl": { "type": ["string", "null"] },
              "scanUrl": { "type": ["string", "null"] },
              "rekorIndex": { "type": ["string", "null"] },
              "rekorUrl": { "type": ["string", "null"] },
              "signerIdentity": { "type": ["string", "null"] },
              "signerIssuer": { "type": ["string", "null"] }
            }
          }
        ]
      },
...
      "WhoamiResult": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": ["string", "null"] },
          "email": { "type": ["string", "null"] },
          "image": { "type": ["string", "null"] },
          "role": { "type": "string", "enum": ["user", "admin"] },
          "authSource": { "type": "string", "enum": ["session", "apikey"] },
          "apiKeyLabel": { "type": "string" }
        }
      },
      "CveCounts": {
        "type": "object",
        "required": ["critical", "high", "medium", "low"],
        "properties": {
          "critical": { "type": "integer" },
          "high": { "type": "integer" },
          "medium": { "type": "integer" },
          "low": { "type": "integer" }
        }
      },
      "ApiError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {}
            }
          }
        }
      }
    }
  }
}